refactor(activerecord): inline MigrationProxy back into deprecator.ts (no re-exports)
