+++ eslint.config.mjs
@@ -86,7 +86,7 @@ export default defineConfig(
       "packages/activerecord/src/encryption/context.ts",
       "packages/activerecord/src/connection-handling.ts",
       // MigrationProxy uses createRequire for synchronous file loading — Node-only
-      "packages/activerecord/src/migration/proxy.ts",
+      "packages/activerecord/src/deprecator.ts",
     ],
     rules: {
       "blazetrails/no-node-builtins": "error",
